+++ src/scooze/utils.py
@@ -138,7 +138,6 @@ def max_card_quantity(fmt: Format) -> int:
             | Format.COMMANDER
             | Format.DUEL
             | Format.GLADIATOR
-            | Format.HISTORICBRAWL
             | Format.OATHBREAKER
             | Format.PAUPERCOMMANDER
             | Format.PREDH
@@ -198,7 +197,7 @@ def main_size(fmt: Format) -> tuple[int, int]:
         ):
             return 60, maxsize
 
-        case Format.BRAWL | Format.HISTORICBRAWL | Format.PAUPERCOMMANDER | Format.PREDH | Format.STANDARDBRAWL:
+        case Format.BRAWL | Format.PAUPERCOMMANDER | Format.PREDH | Format.STANDARDBRAWL:
             return 99, 99
 
         case Format.COMMANDER | Format.DUEL:
@@ -243,7 +242,6 @@ def side_size(fmt: Format) -> tuple[int, int]:
             | Format.COMMANDER
             | Format.DUEL
             | Format.GLADIATOR
-            | Format.HISTORICBRAWL
             | Format.OATHBREAKER
             | Format.PAUPERCOMMANDER
             | Format.PREDH
@@ -281,7 +279,7 @@ def cmdr_size(fmt: Format) -> tuple[int, int]:
         ):
             return 0, 0
 
-        case Format.BRAWL | Format.HISTORICBRAWL | Format.PAUPERCOMMANDER | Format.PREDH | Format.STANDARDBRAWL:
+        case Format.BRAWL | Format.PAUPERCOMMANDER | Format.PREDH | Format.STANDARDBRAWL:
             return 1, 1
 
         case Format.COMMANDER | Format.DUEL:
@@ -323,7 +321,6 @@ def attractions_size(fmt: Format) -> tuple[int, int]:
             | Format.FUTURE
             | Format.GLADIATOR
             | Format.HISTORIC
-            | Format.HISTORICBRAWL
             | Format.MODERN
             | Format.OLDSCHOOL
             | Format.PENNY
@@ -372,7 +369,6 @@ def stickers_size(fmt: Format) -> tuple[int, int]:
             | Format.FUTURE
             | Format.GLADIATOR
             | Format.HISTORIC
-            | Format.HISTORICBRAWL
             | Format.MODERN
             | Format.OLDSCHOOL
             | Format.PENNY